Update Project_A_Gerardo_Sainz.cpp
Did many changes to get the external file data to correctly load to the database, found out the first few lines of the document did mess up the format for how the code reads the data so removed it. Using the addmovie function for the load data function was hard to figure out so just made it so the load data functions puts the data into the array itself instead. I also added the search movie function for option 2, might change the order of my options for my final commit.